requirements.txt
asgiref==3.7.2
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.0
django-allauth==65.3.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.4.0
drf-writable-nested==0.7.0
fastcore==1.5.29
idna==3.10
mysqlclient==2.2.1
packaging==23.2
PyJWT==2.10.1
pytz==2023.3.post1
razorpay==1.4.2
requests==2.31.0
sendmail==2.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.3.0
wfastcgi==3.0.0
